FloorSwitchButtonのファイル移行
